postgress working and so is google api

+++ db/migrate/20160201002835_add_speed_direction_and_gust_to_wind.rb
@@ -1,7 +1,7 @@
 class AddSpeedDirectionAndGustToWind < ActiveRecord::Migration
   def change
-    add_column :winds, :speed, :integer
+    add_column :winds, :speed, :decimal
     add_column :winds, :direction, :string
-    add_column :winds, :gust, :integer
+    add_column :winds, :gust, :decimal
   end
 end
