+++ db/migrate/20160201004504_remove_speed_and_gust_from_winds.rb
@@ -1,6 +1,6 @@
 class RemoveSpeedAndGustFromWinds < ActiveRecord::Migration
   def change
-    remove_column :winds, :speed, :integer
-    remove_column :winds, :gust, :integer
+    remove_column :winds, :speed, :decimal
+    remove_column :winds, :gust, :decimal
   end
 end
